Allow Rails helpers to be called in posts.
The big win here is being able to call asset helpers like
`image_tag`, so that you can keep images within the asset pipeline
and be confident that they will work both in development, and
production.

For example, image URLs in the feed need to be absolute, but the
host is unlikely to be the same in development and production, so
being able to call `<%= image_tag %>` and have it respect the actual
asset host is really useful.

The default handler when rendering inline is `ERB`, and I suspect
that it's not going to be pragmatic to support whitespace-sensitive
handlers like HAML in this way, unfortunately.